+++ .gitignore
@@ -68,5 +68,4 @@ rusty-tags.vi
 costs-*.txt
 names-to-stats.txt
 data_dump_*.bin
-
 .venv
