+++ Cargo.toml
@@ -161,11 +161,7 @@ criterion = { version = "0.5.1", default-features = false, features = [
 crossbeam = "0.8"
 crossbeam-channel = "0.5.8"
 csv = "1.2.1"
-curve25519-dalek = { version = "4.1.3", default-features = false, features = [
-    "alloc",
-    "precomputed-tables",
-    "rand_core",
-] }
+curve25519-dalek = { version = "4.1.3", default-features = false }
 derive-enum-from-into = "0.1.1"
 derive_more = "0.99.9"
 derive-where = "1.2.7"
@@ -173,10 +169,7 @@ dirs = "4"
 dynasm = "2.0"
 dynasmrt = "2.0"
 easy-ext = "0.2"
-ed25519-dalek = { version = "2.1.0", default-features = false, features = [
-    "hazmat",
-    "rand_core",
-] }
+ed25519-dalek = { version = "2.1.0", default-features = false }
 enum-map = "2.1.0"
 enumset = "1.0"
 ethabi = "18"
@@ -334,7 +327,7 @@ rusqlite = { version = "0.29.0", features = ["bundled", "chrono", "functions"] }
 rustc-demangle = "0.1"
 rust-s3 = { version = "0.32.3", features = ["blocking"] }
 rustix = "0.38"
-secp256k1 = { version = "0.27.0", features = ["recovery", "rand-std"] }
+secp256k1 = { version = "0.27.0", default-features = false }
 semver = "1.0.4"
 serde = { version = "1.0.136", features = ["alloc", "derive", "rc"] }
 serde_ignored = "0.1"
@@ -359,7 +352,7 @@ testlib = { path = "test-utils/testlib" }
 test-log = { version = "0.2", default-features = false, features = ["trace"] }
 thiserror = "1.0.30"
 tikv-jemallocator = "0.5.0"
-time = { version = "0.3.9", features = ["parsing", "serde"] }
+time = { version = "0.3.9", default-features = false }
 tokio = { version = "1.28", default-features = false }
 tokio-stream = { version = "0.1.2", features = ["net"] }
 tokio-util = { version = "0.7.1", features = ["codec", "io"] }
