+++ core/crypto/Cargo.toml
@@ -15,22 +15,17 @@ workspace = true
 blake2.workspace = true
 borsh.workspace = true
 bs58.workspace = true
-curve25519-dalek = { version = "4.1", default-features = false, features = [
+curve25519-dalek = { workspace = true, features = [
     "precomputed-tables",
     "alloc",
 ] }
 derive_more.workspace = true
-ed25519-dalek = { version = "2.1", default-features = false, features = [
-    "hazmat",
-] }
+ed25519-dalek = { workspace = true, features = ["hazmat"] }
 hex.workspace = true
 near-account-id.workspace = true
 once_cell.workspace = true
 primitive-types.workspace = true
-secp256k1 = { version = "0.27", default-features = false, features = [
-    "recovery",
-    "alloc",
-] }
+secp256k1 = { workspace = true, features = ["recovery", "alloc"] }
 serde.workspace = true
 serde_json.workspace = true
 stdx.workspace = true
