+++ core/time/Cargo.toml
@@ -12,10 +12,7 @@ publish = true
 workspace = true
 
 [dependencies]
-time = { version = "0.3", features = [
-    "formatting",
-    "parsing",
-], default-features = false }
+time = { workspace = true, features = ["formatting", "parsing"] }
 once_cell = { workspace = true, optional = true }
 tokio = { workspace = true, features = ["time", "sync"], optional = true }
 serde = { workspace = true, optional = true }
@@ -26,4 +23,4 @@ serde_json.workspace = true
 [features]
 default = ["clock", "serde"]
 clock = ["tokio", "once_cell"]
-serde = ["dep:serde"]
+serde = ["dep:serde", "time/serde"]
